Smart 18650 Li-ion Charger BC847
Smart 18650 Li-ion Charger BC847
BC847  

BC847 is in SMD Transistor Series and in SOT-23 sheath. BC847 is an SMD NPN type Transistor. The BC847 Transistor has a collector current of 100mA and a Collector-Emitter voltage of 45V.

 

WHEN THE LI-ON BATTERY IS FULLY FULL, THE GREEN LED WILL BUT ON, WHEN IT IS NOT FULL, the RED LED will turn ON.

 



COMPANENTS:

-BC847

-220R RESİSTOR X2

-1K RESİSTOR

-10K POT

-1N4007 DİODE

-RED AND GREEN LED